feat(#33): 로그아웃 API 액세스 토큰 유효성 검사 로직 추가

+++ src/main/java/friendy/community/domain/auth/service/AuthService.java
@@ -30,7 +30,8 @@ public class AuthService {
         return TokenResponse.of(accessToken, refreshToken);
     }
 
-    public void logout(final String refreshToken) {
+    public void logout(final String accessToken, final String refreshToken) {
+        jwtTokenProvider.validateAccessToken(accessToken);
         final String email = jwtTokenProvider.extractEmailFromRefreshToken(refreshToken);
 
         jwtTokenProvider.deleteRefreshToken(email);
